dataset -> dataframe change, phase 2
