+++ packages/vaex-core/vaex/dataset.py
@@ -1,1 +1,6 @@
-from .dataset import *
+from .dataframe import *
+# alias kept for backward compatibility
+Dataset = DataFrame
+DatasetLocal = DataFrameLocal
+DatasetArrays = DataFrameArrays
+DatasetConcatenated = DataFrameConcatenated
